+++ CHANGELOG.md
@@ -38,6 +38,13 @@
 * No results return when you conduct a Contributor search. Fixes UIIN-2191.
 * Browse contributors | Second pane header doesn't update when user return to "Browse inventory" pane via the web-browser "Back" button. Fixes UIIN-2181.
 * Define new route for Inventory "Browse" page. Refs UIIN-2193.
+* Do not include `highlightMatch` parameter in search queries. Fixes UIIN-2008.
+* Relabel 'Bound-with titles' accordion on item view. Fixes UIIN-2196.
+* Link from bound-with table HRID column to instances. Fixes UIIN-2195.
+* Browse contributors | Show an indicator that contributor value is controlled/linked to an authority record. Refs UIIN-2179.
+* View Instance record | Contributors accordion | Show authorized indicator for each contributor. Refs UIIN-2180.
+* Move Inventory `Browse` logic into separate page: filters and result. Refs UIIN-2194.
+* Ignore .vscode in .gitignore.
 * The exact match result is NOT clickable on the browse result list (Contributors/Call numbers/ Subjects). Refs UIIN-2199.
 
 ## [9.1.0](https://github.com/folio-org/ui-inventory/tree/v9.1.0) (2022-06-28)